Refactor image optimization code
